# bff/openapi.yaml
...
openapi: 3.1.0
info:
  title: Uchiyomi API
  version: 0.32.0
  description: "Everything the web app does, it does over this API. Session JWTs are for the browser; scripts
    use a personal API token (Profile -> Account -> API tokens), sent as `Authorization: Bearer uy_...`.
    Scopes: read (every token; GET plus the library search POST, which is a query), write (anything that changes data), admin (the /api/admin surface, which also
    needs an admin account). /img/* takes the yomi_img cookie, an API token as a Bearer, or the OPDS token
    over HTTP Basic -- so a third-party client needs ONE credential for the JSON and the pictures alike;
    /opds/* takes the OPDS token over HTTP Basic only. List endpoints answer { content: [...] };
    errors answer a non-2xx status with { error, message }. The prose guide with worked examples is
    docs/api.md."
servers:
  - url: /
    description: This server
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: A session JWT from POST /auth/login, or a personal API token (uy_...). Admin routes need the
        admin scope and an admin account.
    opdsBasic:
      type: http
      scheme: basic
      description: HTTP Basic where the password is the OPDS token from POST /api/opds/token.
    imgCookie:
      type: apiKey
      in: cookie
      name: yomi_img
      description: Set at login so <img> tags can authenticate.
...
paths:
...
  /api/books/{id}/download-manifest:
    get:
      tags:
        - downloads
      summary: What the client must fetch to keep a chapter offline.
      operationId: getApiBooksIdDownloadManifest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
        "404":
          description: Not found, or not visible to this account
  /api/downloads:
    get:
      tags:
        - downloads
      summary: The caller's offline queue.
      operationId: getApiDownloads
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
    post:
      tags:
        - downloads
      summary: Queue a chapter for offline use.
      operationId: postApiDownloads
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                bookId:
                  type: string
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "400":
          description: Malformed body or parameter
        "401":
          description: Not authenticated
  /api/downloads/{bookId}:
    delete:
      tags:
        - downloads
      summary: Remove a chapter from the offline queue.
      operationId: deleteApiDownloadsBookid
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
        "404":
          description: Not found, or not visible to this account
...